+++ Expedition/__init__.py
@@ -1,3 +1,4 @@
 from .dll_wrapper import *  # noqa: F403
 from .enums import *  # noqa: F403
+from .exval import ExVal  # noqa: F401
 from .version import __version__ as __version__